+++ src/module/generatemodule.ts
@@ -2,10 +2,9 @@ import { UserInputModel } from '../data/model/userinputmodel';
 import * as vscode from 'vscode';
 import * as mkdirp from 'mkdirp';
 import fs = require('fs');
-import { camelize, pascalize } from '../utils/stringutils';
+import { getPackageName, pascalize } from '../utils/stringutils';
 import { getUserInput } from '../utils/getuserinputformodule';
 import { foldersConstant } from '../data/constant/folders';
-import * as yaml from "js-yaml";
 
 
 export async function generateModule(inUri: vscode.Uri | undefined) {
@@ -28,10 +27,24 @@ async function generateModuleFolders(uri: vscode.Uri, name: string, init: boolea
   if (!paths.includes('lib') || !paths.includes('module')) {
     return vscode.window.showErrorMessage("Please create under lib/module");
   }
+
+  let packageName;
+
+  try {
+    packageName = getPackageName();
+  } catch (error) {
+    return vscode.window.showErrorMessage(error);
+  }
+
+  if (packageName === undefined) {
+    return vscode.window.showErrorMessage("Cant find package name");
+  }
+
   const nameWithoutSpace = name.replace(" ", "_");
   // getPackageNameAndPath();
   const d = uri.path + "/" + nameWithoutSpace;
   await mkdirp(d);
+
 
   const options = { flag: 'wx' };
 
@@ -45,7 +58,7 @@ async function generateModuleFolders(uri: vscode.Uri, name: string, init: boolea
     if (data === '/controller') {
       const newpaths = newp.split("/");
       const libIndex = newpaths.indexOf("lib");
-      const repoImportPath = `'package:${newpaths[libIndex - 1]}/${newpaths.slice(libIndex + 1, newpaths.length - 1).join("/")}/data/repo/${nameWithoutSpace}_repo.dart'`;
+      const repoImportPath = `'package:${packageName}/${newpaths.slice(libIndex + 1, newpaths.length - 1).join("/")}/data/repo/${nameWithoutSpace}_repo.dart'`;
       const file = `${newp}/${nameWithoutSpace}_controller.dart`;
       fs.writeFileSync(file, controllerContent(name, repoImportPath), options);
     }
@@ -54,7 +67,7 @@ async function generateModuleFolders(uri: vscode.Uri, name: string, init: boolea
       const file = `${newp}/${nameWithoutSpace}_screen.dart`;
       const newpaths = newp.split("/");
       const libIndex = newpaths.indexOf("lib");
-      const controllerImportPath = `'package:${newpaths[libIndex - 1]}/${newpaths.slice(libIndex + 1, newpaths.length - 1).join("/")}/controller/${nameWithoutSpace}_controller.dart'`;
+      const controllerImportPath = `'package:${packageName}/${newpaths.slice(libIndex + 1, newpaths.length - 1).join("/")}/controller/${nameWithoutSpace}_controller.dart'`;
       fs.writeFileSync(file, screenContent(name, init, controllerImportPath), options);
     }
 
@@ -169,18 +182,3 @@ class ${widget}Screen extends StatelessWidget {
   return init ? contentWithInit : content;
 }
 
-
-function getPackageNameAndPath(inUri: vscode.Uri) {
-  if (vscode.workspace.workspaceFolders !== undefined) {
-    try {
-      const a = `import 'package:${vscode.workspace.workspaceFolders[0].name}/`;
-
-    } catch (error) {
-      console.log(error);
-    }
-
-  } else {
-    throw new Error("Please open a workspace");
-  }
-
-}