+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # Change Log
 
-All notable changes to the "mnc-flutter-code-standard" extension will be documented in this file.
+## 0.2.0
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+- Stable Module Generator
+- Stable Widget Creator
 
 ## [Unreleased]